third_party/WebKit/LayoutTests/svg/animations/cyclic-syncbase-expected.txt

SVG 1.1 dynamic animation tests

Test cyclic for svg animations for syncbases

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is 100
PASS rootSVGElement.getBBox().y is 0
PASS rootSVGElement.getBBox().x is 0
PASS rootSVGElement.getBBox().y is 100
PASS rootSVGElement.getBBox().x is 100
PASS rootSVGElement.getBBox().y is 0
PASS rootSVGElement.getBBox().x is 0
PASS rootSVGElement.getBBox().y is 100
PASS rootSVGElement.getBBox().x is 100
PASS rootSVGElement.getBBox().y is 0
PASS successfullyParsed is true

TEST COMPLETE
